Remove overly complex enum case sensitivity functionality and rearrange code, removing helpers.rs
